requirements.txt
torch
torchvision
torchmetrics
scikit-learn
numpy
pandas
matplotlib
tqdm
opencv-python
gdown
openpyxl
pyyaml